+++ packages/core/integration-tests/test/javascript.js
@@ -4124,11 +4124,11 @@ describe('javascript', function () {
     assert(
       cssBundleContent.startsWith(
         `body {
-  background-color: #000000;
+  background-color: #000;
 }
 
 .svg-img {
-  background-image: url("data:image/svg+xml,%3Csvg%3E%0A%0A%3C%2Fsvg%3E%0A");
+  background-image: url(data:image/svg+xml,%3Csvg%3E%0A%0A%3C%2Fsvg%3E%0A);
 }`,
       ),
     );
@@ -4166,11 +4166,11 @@ describe('javascript', function () {
     assert(
       cssBundleContent.startsWith(
         `body {
-  background-color: #000000;
+  background-color: #000;
 }
 
 .svg-img {
-  background-image: url("data:image/svg+xml,%3Csvg%3E%0A%0A%3C%2Fsvg%3E%0A");
+  background-image: url(data:image/svg+xml,%3Csvg%3E%0A%0A%3C%2Fsvg%3E%0A);
 }`,
       ),
     );
@@ -4254,11 +4254,11 @@ describe('javascript', function () {
     assert(
       cssBundleContent.startsWith(
         `body {
-  background-color: #000000;
+  background-color: #000;
 }
 
 .svg-img {
-  background-image: url("data:image/svg+xml,%3Csvg%3E%0A%0A%3C%2Fsvg%3E%0A");
+  background-image: url(data:image/svg+xml,%3Csvg%3E%0A%0A%3C%2Fsvg%3E%0A);
 }`,
       ),
     );