+++ packages/transformers/css/src/CSSTransformer.js
@@ -1,18 +1,25 @@
 // @flow
 
 import type {Root} from 'postcss';
-import type {FilePath} from '@parcel/types';
+import type {FilePath, MutableAsset} from '@parcel/types';
 
+import {hashString} from '@parcel/hash';
 import SourceMap from '@parcel/source-map';
 import {Transformer} from '@parcel/plugin';
 import {createDependencyLocation, remapSourceLocation} from '@parcel/utils';
 import postcss from 'postcss';
 import nullthrows from 'nullthrows';
 import valueParser from 'postcss-value-parser';
+import postcssModules from 'postcss-modules';
+import FileSystemLoader from 'css-modules-loader-core/lib/file-system-loader';
 import semver from 'semver';
+import path from 'path';
 
 const URL_RE = /url\s*\("?(?![a-z]+:)/;
 const IMPORT_RE = /@import/;
+const COMPOSES_RE = /composes:.+from\s*("|').*("|')\s*;?/;
+const FROM_IMPORT_RE = /.+from\s*(?:"|')(.*)(?:"|')\s*;?/;
+const MODULE_BY_NAME_RE = /\.module\./;
 
 function canHaveDependencies(filePath: FilePath, code: string) {
   return !/\.css$/.test(filePath) || IMPORT_RE.test(code) || URL_RE.test(code);
@@ -30,12 +37,19 @@ export default (new Transformer({
     // to be filled in later. When the CSS transformer runs, it would pick that up and try to
     // resolve a dependency for the id which obviously doesn't exist. Also, it's faster to do
     // it this way since the resulting CSS doesn't need to be re-parsed.
-    if (asset.meta.hasDependencies === false) {
+    let isCSSModule =
+      asset.meta.cssModulesCompiled !== true &&
+      MODULE_BY_NAME_RE.test(asset.filePath);
+    if (asset.meta.hasDependencies === false && !isCSSModule) {
       return null;
     }
 
     let code = await asset.getCode();
-    if (code != null && !canHaveDependencies(asset.filePath, code)) {
+    if (
+      code != null &&
+      !canHaveDependencies(asset.filePath, code) &&
+      !isCSSModule
+    ) {
       return null;
     }
 
@@ -50,9 +64,10 @@ export default (new Transformer({
     };
   },
 
-  async transform({asset}) {
+  async transform({asset, resolve, options}) {
     // Normalize the asset's environment so that properties that only affect JS don't cause CSS to be duplicated.
     // For example, with ESModule and CommonJS targets, only a single shared CSS bundle should be produced.
+    let env = asset.env;
     asset.setEnvironment({
       context: 'browser',
       engines: {
@@ -62,22 +77,36 @@ export default (new Transformer({
       sourceMap: asset.env.sourceMap,
     });
 
+    let isCSSModule =
+      asset.meta.cssModulesCompiled !== true &&
+      MODULE_BY_NAME_RE.test(asset.filePath);
+
     // Check for `hasDependencies` being false here as well, as it's possible
     // another transformer (such as PostCSSTransformer) has already parsed an
     // ast and CSSTransformer's parse was never called.
     let ast = await asset.getAST();
-    if (!ast || asset.meta.hasDependencies === false) {
+    if (!ast || (asset.meta.hasDependencies === false && !isCSSModule)) {
       return [asset];
     }
 
     let program: Root = postcss.fromJSON(ast.program);
+    let assets = [asset];
+    if (isCSSModule) {
+      assets = await compileCSSModules(asset, env, program, resolve, options);
+    }
+
+    if (asset.meta.hasDependencies === false) {
+      return assets;
+    }
+
     let originalSourceMap = await asset.getMap();
-    let createLoc = (start, specifier, lineOffset, colOffset) => {
+    let createLoc = (start, specifier, lineOffset, colOffset, o) => {
       let loc = createDependencyLocation(
         start,
         specifier,
         lineOffset,
         colOffset,
+        o,
       );
       if (originalSourceMap) {
         loc = remapSourceLocation(loc, originalSourceMap);
@@ -142,19 +171,22 @@ export default (new Transformer({
             node.nodes.length > 0 &&
             !node.nodes[0].value.startsWith('#') // IE's `behavior: url(#default#VML)`
           ) {
-            let url = asset.addURLDependency(node.nodes[0].value, {
+            let urlNode = node.nodes[0];
+            let url = asset.addURLDependency(urlNode.value, {
               loc:
                 decl.source &&
                 decl.source.start &&
                 createLoc(
                   decl.source.start,
-                  node.nodes[0].value,
+                  urlNode.value,
                   0,
-                  node.nodes[0].sourceIndex,
+                  decl.source.start.offset + urlNode.sourceIndex + 1,
+                  0,
                 ),
             });
-            isDeclDirty = node.nodes[0].value !== url;
-            node.nodes[0].value = url;
+            isDeclDirty = urlNode.value !== url;
+            urlNode.type = 'word';
+            urlNode.value = url;
           }
         });
 
@@ -172,7 +204,7 @@ export default (new Transformer({
       });
     }
 
-    return [asset];
+    return assets;
   },
 
   async generate({asset, ast, options}) {
@@ -207,3 +239,124 @@ export default (new Transformer({
     };
   },
 }): Transformer);
+
+async function compileCSSModules(asset, env, program, resolve, options) {
+  let cssModules;
+
+  let code = asset.isASTDirty() ? null : await asset.getCode();
+  if (code == null || COMPOSES_RE.test(code)) {
+    program.walkDecls(decl => {
+      let [, importPath] = FROM_IMPORT_RE.exec(decl.value) || [];
+      if (decl.prop === 'composes' && importPath != null) {
+        let parsed = valueParser(decl.value);
+        let start = (decl.source.start: any);
+
+        parsed.walk(node => {
+          if (node.type === 'string') {
+            asset.addDependency({
+              specifier: importPath,
+              specifierType: 'url',
+              loc: start
+                ? {
+                    filePath: asset.filePath,
+                    start,
+                    end: {
+                      line: start.line,
+                      column: start.column + importPath.length,
+                    },
+                  }
+                : undefined,
+            });
+          }
+        });
+      }
+    });
+  }
+
+  let {root} = await postcss([
+    postcssModules({
+      getJSON: (filename, json) => (cssModules = json),
+      Loader: createLoader(asset, resolve),
+      generateScopedName: (name, filename) =>
+        `${name}_${hashString(
+          path.relative(options.projectRoot, filename),
+        ).substr(0, 6)}`,
+    }),
+  ]).process(program, {from: asset.filePath, to: asset.filePath});
+  asset.setAST({
+    type: 'postcss',
+    version: '8.2.1',
+    program: root.toJSON(),
+  });
+
+  let assets = [asset];
+  if (cssModules) {
+    // $FlowFixMe
+    let cssModulesList = (Object.entries(cssModules): Array<[string, string]>);
+    let deps = asset.getDependencies().filter(dep => dep.priority === 'sync');
+    let code: string;
+    if (deps.length > 0) {
+      code = `
+        module.exports = Object.assign({}, ${deps
+          .map(dep => `require(${JSON.stringify(dep.specifier)})`)
+          .join(', ')}, ${JSON.stringify(cssModules, null, 2)});
+      `;
+    } else {
+      code = cssModulesList
+        .map(
+          // This syntax enables shaking the invidual statements, so that unused classes don't even exist in JS.
+          ([className, classNameHashed]) =>
+            `module.exports[${JSON.stringify(className)}] = ${JSON.stringify(
+              classNameHashed,
+            )};`,
+        )
+        .join('\n');
+    }
+
+    asset.symbols.ensure();
+    for (let [k, v] of cssModulesList) {
+      asset.symbols.set(k, v);
+    }
+    asset.symbols.set('default', 'default');
+
+    assets.push({
+      type: 'js',
+      content: code,
+      env,
+    });
+  }
+  return assets;
+}
+
+function createLoader(
+  asset: MutableAsset,
+  resolve: (from: FilePath, to: string) => Promise<FilePath>,
+) {
+  return class ParcelFileSystemLoader extends FileSystemLoader {
+    async fetch(composesPath, relativeTo) {
+      let importPath = composesPath.replace(/^["']|["']$/g, '');
+      let resolved = await resolve(relativeTo, importPath);
+      let rootRelativePath = path.resolve(path.dirname(relativeTo), resolved);
+      let root = path.resolve('/');
+      // fixes an issue on windows which is part of the css-modules-loader-core
+      // see https://github.com/css-modules/css-modules-loader-core/issues/230
+      if (rootRelativePath.startsWith(root)) {
+        rootRelativePath = rootRelativePath.substr(root.length);
+      }
+
+      let source = await asset.fs.readFile(resolved, 'utf-8');
+      let {exportTokens} = await this.core.load(
+        source,
+        rootRelativePath,
+        undefined,
+        // $FlowFixMe[method-unbinding]
+        this.fetch.bind(this),
+      );
+      return exportTokens;
+    }
+
+    get finalSource() {
+      return '';
+    }
+  };
+}