+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to Parcel will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and Parcel adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [2.2.0] - 2022-01-12
+
+### Added
+
+- New `@parcel/transformer-css-experimental` plugin, which is powered by [@parcel/css](https://github.com/parcel-bundler/parcel-css) - [Details](https://github.com/parcel-bundler/parcel/pull/7538)
+
+### Fixed
+
+- Updated `node-forge` to 1.0.0 to fix security vulnerability
+
 ## [2.1.1] - 2022-01-06
 
 ### Fixed